split code into files, use webpack to bundle
