+++ package.json
@@ -19,11 +19,13 @@
   },
   "homepage": "https://github.com/Ahmed-Salama/YetAnotherPhysicsEngine#readme",
   "dependencies": {
-    "immutable": "^3.8.2",
-    "typescript": "^2.6.2"
+    "immutable": "^3.8.2"
   },
   "devDependencies": {
     "@types/jquery": "^3.2.17",
-    "@types/rx": "^4.1.1"
+    "@types/rx": "^4.1.1",
+    "ts-loader": "^3.2.0",
+    "typescript": "^2.6.2",
+    "webpack": "^3.10.0"
   }
 }
